Version 2.0
This version features a simpler way to customize the INI parsing phased. As the
INI file format is not really a standard, an easier way to allow the library
user to change the way INI data is parser is a must.
Now you need to pass a configuration object to an IniParser instance, which will be
used to set the behaviour of the parser. A default implementation is used
if none is passed.

For this version the code base organization has been improved, both in the library
project and in the unit test project. New namespaces has been added, and
files are stored physically in folders matching the namespaces paths.

Finally, project files have been migrated to Visual Studio 2010, and
Mono Support has been added.