Refactors: Use TextReader, avoid returning null
TextReader is a parent class of StreamReader and StringReader, so this simplifies the number of Parse methods to add.
When an error is found during Parsing, and the parser is configured to not throw exceptions, now we return an empty
IniData structure instead of null. This options should be safer for input external to the program.